move habitatctl back
