+++ cmd/habitatctl/commands/helpers.go
