+++ cmd/habitatctl/commands/stop.go
