+++ cmd/habitatctl/main.go
@@ -4,7 +4,7 @@ Copyright © 2021 NAME HERE <EMAIL ADDRESS>
 */
 package main
 
-import "github.com/eagraf/habitat/pkg/habitatctl/commands"
+import "github.com/eagraf/habitat/cmd/habitatctl/commands"
 
 func main() {
 	commands.Execute()
